+++ src/snowflake/snowpark/modin/plugin/extensions/index.py
@@ -23,7 +23,6 @@
 
 from __future__ import annotations
 
-from functools import wraps
 from typing import Any, Callable, Hashable, Iterator, Literal
 
 import modin
@@ -49,63 +48,6 @@ from snowflake.snowpark.modin.plugin.utils.error_message import (
 from snowflake.snowpark.modin.plugin.utils.warning_message import WarningMessage
 
 
-def is_lazy_check(func: Callable) -> Callable:
-    """
-    Decorator method for separating function calls for lazy indexes and non-lazy (column) indexes
-    """
-
-    @wraps(func)
-    def check_lazy(*args: Any, **kwargs: Any) -> Any:
-        func_name = func.__name__
-
-        # If the index is lazy, call the method and return
-        if args[0].is_lazy:
-            returned_value = func(*args, **kwargs)
-            return returned_value
-        else:
-            # If the index is not lazy, get the cached native index and call the function
-            native_index = args[0]._index
-            native_func = getattr(native_index, func_name)
-
-            # If the function is a property, we will get a non-callable, so we just return it
-            # Examples of this are values or dtype
-            if not callable(native_func):
-                return native_func
-
-            # Remove the first argument in args, because it is `self` and we don't need it
-            args = args[1:]
-            args = tuple(try_convert_index_to_native(a) for a in args)
-            for k, v in kwargs.items():
-                kwargs[k] = try_convert_index_to_native(v)
-            returned_value = native_func(*args, **kwargs)
-
-            # If we return a native Index, we need to convert this to a modin index but keep it locally.
-            # Examples of this are `astype` and `copy`
-            if isinstance(returned_value, native_pd.Index):
-                returned_value = Index(returned_value, convert_to_lazy=False)
-            # Some methods also return a tuple with a pandas Index, so convert the tuple's first item to a modin Index
-            # Examples of this are `_get_indexer_strict` and `sort_values`
-            elif isinstance(returned_value, tuple) and isinstance(
-                returned_value[0], native_pd.Index
-            ):
-                returned_value = (
-                    Index(returned_value[0], convert_to_lazy=False),
-                    returned_value[1],
-                )
-            # For methods that return a series, convert this series to snowpark pandas
-            # an example is to_series
-            elif isinstance(returned_value, native_pd.Series):
-                returned_value = Series(returned_value)
-
-            # for methods that return a dataframe, convert this dataframe to snowpark pandas
-            elif isinstance(returned_value, native_pd.DataFrame):
-                returned_value = DataFrame(returned_value)
-
-            return returned_value
-
-    return check_lazy
-
-
 class Index(metaclass=TelemetryMeta):
     def __init__(
         self,
@@ -114,7 +56,6 @@ class Index(metaclass=TelemetryMeta):
         copy: bool = False,
         name: object = None,
         tupleize_cols: bool = True,
-        convert_to_lazy: bool = True,
     ) -> None:
         """
         Immutable sequence used for indexing and alignment.
@@ -134,9 +75,6 @@ class Index(metaclass=TelemetryMeta):
             Name to be stored in the index.
         tupleize_cols : bool (default: True)
             When True, attempt to create a MultiIndex if possible.
-        convert_to_lazy : bool (default: True)
-            When True, create a lazy index object from a local data input, otherwise, create an index object that saves a pandas index locally.
-            We only set convert_to_lazy as False to avoid pulling data back and forth from Snowflake, e.g., when calling df.columns, the column data should always be kept locally.
 
         Notes
         -----
@@ -155,35 +93,6 @@ class Index(metaclass=TelemetryMeta):
         >>> pd.Index([1, 2, 3], dtype="uint8")
         Index([1, 2, 3], dtype='int64')
         """
-        self.is_lazy = convert_to_lazy
-        if self.is_lazy:
-            self.set_query_compiler(
-                data=data,
-                dtype=dtype,
-                copy=copy,
-                name=name,
-                tupleize_cols=tupleize_cols,
-            )
-        else:
-            self.set_local_index(
-                data=data,
-                dtype=dtype,
-                copy=copy,
-                name=name,
-                tupleize_cols=tupleize_cols,
-            )
-
-    def set_query_compiler(
-        self,
-        data: ArrayLike | SnowflakeQueryCompiler | None = None,
-        dtype: str | np.dtype | ExtensionDtype | None = None,
-        copy: bool = False,
-        name: object = None,
-        tupleize_cols: bool = True,
-    ) -> None:
-        """
-        Helper method to find and save query compiler when index should be lazy
-        """
         if isinstance(data, SnowflakeQueryCompiler):
             qc = data
         else:
@@ -197,29 +106,6 @@ class Index(metaclass=TelemetryMeta):
                 )
             )._query_compiler
         self._query_compiler = qc.drop(columns=qc.columns)
-
-    def set_local_index(
-        self,
-        data: ArrayLike | SnowflakeQueryCompiler | None = None,
-        dtype: str | np.dtype | ExtensionDtype | None = None,
-        copy: bool = False,
-        name: object = None,
-        tupleize_cols: bool = True,
-    ) -> None:
-        """
-        Helper method to create and save local index when index should not be lazy
-        """
-        if isinstance(data, SnowflakeQueryCompiler):
-            index = data._modin_frame.index_columns_pandas_index()
-        else:
-            index = native_pd.Index(
-                data=data,
-                dtype=dtype,
-                copy=copy,
-                name=name,
-                tupleize_cols=tupleize_cols,
-            )
-        self._index = index
 
     def __getattr__(self, key: str) -> Any:
         """
@@ -268,11 +154,9 @@ class Index(metaclass=TelemetryMeta):
             pandas Index
                 A native pandas Index representation of self
         """
-        if self.is_lazy:
-            return self._query_compiler._modin_frame.index_columns_pandas_index(
-                statement_params=statement_params, **kwargs
-            )
-        return self._index
+        return self._query_compiler._modin_frame.index_columns_pandas_index(
+            statement_params=statement_params, **kwargs
+        )
 
     @property
     def values(self) -> ArrayLike:
@@ -368,12 +252,9 @@ class Index(metaclass=TelemetryMeta):
         >>> idx.is_unique
         True
         """
-        if not self.is_lazy:
-            return self._index.is_unique
         return self._query_compiler._modin_frame.has_unique_index()
 
     @property
-    @is_lazy_check
     def has_duplicates(self) -> bool:
         """
         Check if the Index has duplicate values.
@@ -409,7 +290,6 @@ class Index(metaclass=TelemetryMeta):
         """
         return not self.is_unique
 
-    @is_lazy_check
     def unique(self, level: Hashable | None = None) -> Index:
         """
         Return unique values in the index.
@@ -453,7 +333,6 @@ class Index(metaclass=TelemetryMeta):
         )
 
     @property
-    @is_lazy_check
     def dtype(self) -> DtypeObj:
         """
         Get the dtype object of the underlying data.
@@ -493,7 +372,6 @@ class Index(metaclass=TelemetryMeta):
         """
         return (len(self),)
 
-    @is_lazy_check
     def astype(self, dtype: str | type | ExtensionDtype, copy: bool = True) -> Index:
         """
         Create an Index with values cast to dtypes.
@@ -564,19 +442,13 @@ class Index(metaclass=TelemetryMeta):
         """
         Set Index name.
         """
-        if self.is_lazy:
-            self._query_compiler = self._query_compiler.set_index_names([value])
-        else:
-            self._index.name = value
+        self._query_compiler = self._query_compiler.set_index_names([value])
 
     def _get_names(self) -> list[Hashable]:
         """
         Get names of index
         """
-        if self.is_lazy:
-            return self._query_compiler.get_index_names()
-        else:
-            return self.to_pandas().names
+        return self._query_compiler.get_index_names()
 
     def _set_names(self, values: list) -> None:
         """
@@ -591,10 +463,7 @@ class Index(metaclass=TelemetryMeta):
         ------
         TypeError if each name is not hashable.
         """
-        if self.is_lazy:
-            self._query_compiler = self._query_compiler.set_index_names(values)
-        else:
-            self._index.names = values
+        self._query_compiler = self._query_compiler.set_index_names(values)
 
     names = property(fset=_set_names, fget=_get_names)
 
@@ -633,8 +502,7 @@ class Index(metaclass=TelemetryMeta):
         WarningMessage.index_to_pandas_warning("set_names")
         if not inplace:
             return Index(
-                self.to_pandas().set_names(names, level=level, inplace=inplace),
-                convert_to_lazy=self.is_lazy,
+                self.to_pandas().set_names(names, level=level, inplace=inplace)
             )
         return self.to_pandas().set_names(names, level=level, inplace=inplace)
 
@@ -666,7 +534,6 @@ class Index(metaclass=TelemetryMeta):
         return len(self)
 
     @property
-    @is_lazy_check
     def nlevels(self) -> int:
         """
         Number of levels.
@@ -863,7 +730,6 @@ class Index(metaclass=TelemetryMeta):
         """
         # TODO: SNOW-1458142 implement argmax
 
-    @is_lazy_check
     def copy(
         self,
         name: Hashable | None = None,
@@ -898,11 +764,7 @@ class Index(metaclass=TelemetryMeta):
         False
         """
         WarningMessage.ignored_argument(operation="copy", argument="deep", message="")
-        return Index(
-            self._query_compiler.copy(),
-            name=name,
-            convert_to_lazy=self.is_lazy,
-        )
+        return Index(self._query_compiler.copy(), name=name)
 
     @index_not_implemented()
     def delete(self) -> None:
@@ -926,7 +788,6 @@ class Index(metaclass=TelemetryMeta):
         """
         # TODO: SNOW-1458146 implement delete
 
-    @is_lazy_check
     def drop(
         self,
         labels: Any,
@@ -959,10 +820,7 @@ class Index(metaclass=TelemetryMeta):
         """
         # TODO: SNOW-1458146 implement drop
         WarningMessage.index_to_pandas_warning("drop")
-        return Index(
-            self.to_pandas().drop(labels=labels, errors=errors),
-            convert_to_lazy=self.is_lazy,
-        )
+        return Index(self.to_pandas().drop(labels=labels, errors=errors))
 
     @index_not_implemented()
     def drop_duplicates(self) -> None:
@@ -988,7 +846,6 @@ class Index(metaclass=TelemetryMeta):
         """
         # TODO: SNOW-1458147 implement drop_duplicates
 
-    @is_lazy_check
     def duplicated(self, keep: Literal["first", "last", False] = "first") -> np.ndarray:
         """
         Indicate duplicate index values.
@@ -1117,20 +974,9 @@ class Index(metaclass=TelemetryMeta):
         if isinstance(other, native_pd.Index):
             # Same as DataFrame/Series equals. Convert native Index to Snowpark pandas
             # Index for comparison.
-            other = Index(other, convert_to_lazy=self.is_lazy)
+            other = Index(other)
 
-        left = self
-        right = other
-        # If both are cached compare underlying cached value locally.
-        if not left.is_lazy and not right.is_lazy:
-            return left._index.equals(right._index)
-
-        # Ensure both sides are lazy before calling index_equals on query_compiler.
-        if not left.is_lazy:
-            left = Index(left._index, convert_to_lazy=True)
-        if not right.is_lazy:
-            right = Index(right._index, convert_to_lazy=True)
-        return left._query_compiler.index_equals(right._query_compiler)
+        return self._query_compiler.index_equals(other._query_compiler)
 
     @index_not_implemented()
     def identical(self) -> None:
@@ -1446,7 +1292,6 @@ class Index(metaclass=TelemetryMeta):
         """
         # TODO: SNOW-1458122 implement rename
 
-    @is_lazy_check
     def nunique(self, dropna: bool = True) -> int:
         """
         Return number of unique elements in the object.
@@ -1483,7 +1328,6 @@ class Index(metaclass=TelemetryMeta):
         """
         return self._query_compiler.nunique_index(dropna=dropna)
 
-    @is_lazy_check
     def value_counts(
         self,
         normalize: bool = False,
@@ -1565,7 +1409,6 @@ class Index(metaclass=TelemetryMeta):
             name="proportion" if normalize else "count",
         )
 
-    @is_lazy_check
     def item(self) -> Hashable:
         """
         Return the first element of the underlying data as a Python scalar.
@@ -1592,7 +1435,6 @@ class Index(metaclass=TelemetryMeta):
         # otherwise raise the same value error as pandas
         raise ValueError("can only convert an array of size 1 to a Python scalar")
 
-    @is_lazy_check
     def to_series(
         self, index: Index | None = None, name: Hashable | None = None
     ) -> Series:
@@ -1636,7 +1478,6 @@ class Index(metaclass=TelemetryMeta):
         ser.name = name
         return ser
 
-    @is_lazy_check
     def to_frame(
         self, index: bool = True, name: Hashable | None = lib.no_default
     ) -> modin.pandas.DataFrame:
@@ -1863,7 +1704,6 @@ class Index(metaclass=TelemetryMeta):
 
     to_list = tolist
 
-    @is_lazy_check
     def sort_values(
         self,
         return_indexer: bool = False,
@@ -1940,7 +1780,7 @@ class Index(metaclass=TelemetryMeta):
             key=key,
             include_indexer=return_indexer,
         )
-        index = Index(res, convert_to_lazy=self.is_lazy)
+        index = Index(res)
         if return_indexer:
             # When `return_indexer` is True, `res` is a query compiler with one index column
             # and one data column.
@@ -1989,7 +1829,6 @@ class Index(metaclass=TelemetryMeta):
         """
         # TODO: SNOW-1458150 implement join
 
-    @is_lazy_check
     def intersection(self, other: Any, sort: bool = False) -> Index:
         """
         Form the intersection of two Index objects.
@@ -2024,11 +1863,9 @@ class Index(metaclass=TelemetryMeta):
         return Index(
             self.to_pandas().intersection(
                 other=try_convert_index_to_native(other), sort=sort
-            ),
-            convert_to_lazy=self.is_lazy,
+            )
         )
 
-    @is_lazy_check
     def union(self, other: Any, sort: bool = False) -> Index:
         """
         Form the union of two Index objects.
@@ -2077,11 +1914,9 @@ class Index(metaclass=TelemetryMeta):
         # TODO: SNOW-1468240 implement union w/ sort
         WarningMessage.index_to_pandas_warning("union")
         return Index(
-            self.to_pandas().union(other=try_convert_index_to_native(other), sort=sort),
-            convert_to_lazy=self.is_lazy,
+            self.to_pandas().union(other=try_convert_index_to_native(other), sort=sort)
         )
 
-    @is_lazy_check
     def difference(self, other: Any, sort: Any = None) -> Index:
         """
         Return a new Index with elements of index not in `other`.
@@ -2118,11 +1953,9 @@ class Index(metaclass=TelemetryMeta):
         # TODO: SNOW-1458152 implement difference
         WarningMessage.index_to_pandas_warning("difference")
         return Index(
-            self.to_pandas().difference(try_convert_index_to_native(other), sort=sort),
-            convert_to_lazy=self.is_lazy,
+            self.to_pandas().difference(try_convert_index_to_native(other), sort=sort)
         )
 
-    @is_lazy_check
     def get_indexer_for(self, target: Any) -> Any:
         """
         Guaranteed return of an indexer even when non-unique.
@@ -2145,16 +1978,14 @@ class Index(metaclass=TelemetryMeta):
         WarningMessage.index_to_pandas_warning("get_indexer_for")
         return self.to_pandas().get_indexer_for(target=target)
 
-    @is_lazy_check
     def _get_indexer_strict(self, key: Any, axis_name: str) -> tuple[Index, np.ndarray]:
         """
         Analogue to pandas.Index.get_indexer that raises if any elements are missing.
         """
         WarningMessage.index_to_pandas_warning("_get_indexer_strict")
         tup = self.to_pandas()._get_indexer_strict(key=key, axis_name=axis_name)
-        return Index(tup[0], convert_to_lazy=self.is_lazy), tup[1]
+        return Index(tup[0]), tup[1]
 
-    @is_lazy_check
     def get_level_values(self, level: int | str) -> Index:
         """
         Return an Index of values for requested level.
@@ -2188,9 +2019,7 @@ class Index(metaclass=TelemetryMeta):
         Index(['a', 'b', 'c'], dtype='object')
         """
         WarningMessage.index_to_pandas_warning("get_level_values")
-        return Index(
-            self.to_pandas().get_level_values(level=level), convert_to_lazy=self.is_lazy
-        )
+        return Index(self.to_pandas().get_level_values(level=level))
 
     @index_not_implemented()
     def isin(self) -> None:
@@ -2233,7 +2062,6 @@ class Index(metaclass=TelemetryMeta):
         """
         # TODO: SNOW-1458153 implement isin
 
-    @is_lazy_check
     def slice_indexer(
         self,
         start: Hashable | None = None,
@@ -2279,14 +2107,12 @@ class Index(metaclass=TelemetryMeta):
         return self.to_pandas().slice_indexer(start=start, end=end, step=step)
 
     @property
-    @is_lazy_check
     def array(self) -> ExtensionArray:
         """
         return the array of values
         """
         return self.to_pandas().array
 
-    @is_lazy_check
     def _summary(self, name: Any = None) -> str:
         """
         Return a summarized representation.
@@ -2304,14 +2130,12 @@ class Index(metaclass=TelemetryMeta):
         WarningMessage.index_to_pandas_warning("_summary")
         return self.to_pandas()._summary(name=name)
 
-    @is_lazy_check
     def __array__(self, dtype: Any = None) -> np.ndarray:
         """
         The array interface, return the values.
         """
         return self.to_pandas().__array__(dtype=dtype)
 
-    @is_lazy_check
     def __repr__(self) -> str:
         """
         Return a string representation for this object.
@@ -2459,7 +2283,6 @@ class Index(metaclass=TelemetryMeta):
 
         return repr
 
-    @is_lazy_check
     def __iter__(self) -> Iterator:
         """
         Return an iterator of the values.
@@ -2485,7 +2308,6 @@ class Index(metaclass=TelemetryMeta):
         WarningMessage.index_to_pandas_warning("__iter__")
         return self.to_pandas().__iter__()
 
-    @is_lazy_check
     def __contains__(self, key: Any) -> bool:
         """
         Return a boolean indicating whether the provided key is in the index.
@@ -2519,14 +2341,12 @@ class Index(metaclass=TelemetryMeta):
         WarningMessage.index_to_pandas_warning("__contains__")
         return self.to_pandas().__contains__(key=key)
 
-    @is_lazy_check
     def __len__(self) -> int:
         """
         Return the length of the Index as an int.
         """
         return self._query_compiler.get_axis_len(0)
 
-    @is_lazy_check
     def __getitem__(self, key: Any) -> np.ndarray | None | Index:
         """
         Reuse series iloc to implement getitem for index.
@@ -2542,7 +2362,6 @@ class Index(metaclass=TelemetryMeta):
                 "boolean arrays are valid indices"
             ) from ie
 
-    @is_lazy_check
     def __setitem__(self, key: Any, value: Any) -> None:
         """
         Override numpy.ndarray's __setitem__ method to work as desired.